gomod: bump lnd to version 0.18.0-beta

+++ sweep/sweeper.go
@@ -217,5 +217,5 @@ func (s *Sweeper) GetSweepFee(ctx context.Context,
 
 	weight := weightEstimate.Weight()
 
-	return feeRate.FeeForWeight(int64(weight)), nil
+	return feeRate.FeeForWeight(weight), nil
 }
